move userTags pagination utils
