+++ src/utilities/userTagsPaginationUtils/index.ts
@@ -1,3 +1,6 @@
 export * from "./parseUserTagSortedBy";
 export * from "./parseUserTagWhere";
 export * from "./parseUserTagMemberWhere";
+export * from "./getUserTagGraphQLConnectionSort";
+export * from "./getUserTagGraphQLConnectionFilter";
+export * from "./getUserTagMemberGraphQLConnectionFilter";
